isolated played games logic
